Refactor file and directory creation and deletion methods

+++ api/src/util/storage.util.js
@@ -39,7 +39,9 @@ module.exports.purge = async () => {
       await Promise.all(promises);
 
       const ids = files.map(({ id }) => id);
-      db.prepare(`DELETE FROM match WHERE id IN (${ids.join(',')})`).run();
+      if (ids.length) {
+        db.prepare(`DELETE FROM match WHERE id IN (${ids.join(',')})`).run();
+      }
 
       if (files.length > 0) console.log(`purged ${files.length} file(s)`);
     } catch (error) {
